requirements.txt
pandas>=1.5.0
numpy>=1.21.0
pgmpy>=0.1.22
scikit-learn>=1.0.0
matplotlib>=3.5.0
torch>=2.0.0
transformers>=4.30.0
tqdm>=4.65.0
huggingface_hub>=0.17.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=1.2.6,<2.0.0
langgraph>=0.0.40
pydantic>=2.0.0
httpx>=0.24.0
python-dotenv>=1.0.0
scipy>=1.9.0
tenacity>=8.2.0
plotly>=5.14.0
